% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_self2neighborhood.R
\name{add_self2neighborhood}
\alias{add_self2neighborhood}
\title{add_self2neighborhood}
\usage{
add_self2neighborhood(NN_ids, i_skip = c())
}
\arguments{
\item{NN_ids}{NN_ids object from locPGSacc()}

\item{i_skip}{(optional) integer vector: list of indices to skip when adding element i to its own NN vector}
}
\value{
Returns inputted NN_ids object but with element i added to the i'th NN vector
}
\description{
Takes an NN_ids object from \code{\link[=locPGSacc]{locPGSacc()}} and adds element i to the i'th NN vector
}
\author{
Nuno R. G. Carvalho: \email{nunocarvalho@gatech.edu}
}
