% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGSdecay.R
\name{bin_dim}
\alias{bin_dim}
\title{bin_dim}
\usage{
bin_dim(data, ref_window = 0.95, bins = 15)
}
\arguments{
\item{ref_window}{(optional) numeric: proportion of the distribution, centered at the median, to include consider when making the non-leftmost and -rightmost bins. Also affects standardization of portability slope}

\item{bins}{(optional) integer: number of bins to split up the middle of the sample by according to the dimensional variable. The leftmost and rightmost bins consist of individuals outside of the central window given by ref_window}
}
\value{
Returns a vector of dim groups to apply directly to the dataset:
}
\description{
splits a complete, renamed dataset by bins along some variable
}
\details{
Takes a dataset with complete cases and with column names standardized
('dim','pheno','PGS') and splits individuals into bins depending on their 'dim'
variable. Internal function.
}
\author{
Nuno R. G. Carvalho: \email{nunocarvalho@gatech.edu}
}
