% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_dist.R
\name{dim_dist}
\alias{dim_dist}
\title{dim_dist}
\usage{
dim_dist(
  data,
  col_dims,
  reference_point = NA,
  centroid_indices = NA,
  col_dist = "dim_dist"
)
}
\arguments{
\item{data}{data table containing all necessary columns and rows}

\item{col_dims}{character vector: name of the dimension column(s)}

\item{reference_point}{(optional) specifies what reference point is used to calculate sample distance from. One of:
\itemize{
\item default: uses centroid (mean position) of all samples. Specify which samples to consider for centroid using 'centroid_indices' parameter.
\item 0: origin of the space
\item numeric vector (same length as col_dims)
}}

\item{centroid_indices}{(optional) integer or logical vector: indices/mask of the data's samples to use for centroid calculation. All samples are used by default}

\item{col_dist}{(optional) character: output column name of sample's distance in space}
}
\description{
Computes each sample's distance from a specified reference point (centroid if nto given)
}
\details{
This function takes a dataset containing sample coordinates in some
dimensional space and computes each sample's distance from a reference point.
If not specified, the reference point will be automatically set to the
centroid (mean position) of all the samples. If specified, the centroid will
only be computed for a subset of samples.
}
\author{
Nuno R. G. Carvalho: \email{nunocarvalho@gatech.edu}
}
