% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locPGSacc.R
\name{get_accuracy}
\alias{get_accuracy}
\title{get_accuracy}
\usage{
get_accuracy(data, NN_ids_i, col_pheno, col_PGS)
}
\arguments{
\item{col_pheno}{character: column name of the phenotype of interest}

\item{col_PGS}{character: column name of the polygenic scores for the phenotype of interest}

\item{NN_ids_i:}{numeric vector: indices in 'data' that are part of neighborhood to compute correlation within}
}
\value{
Returns correlation value between phenotype and PGS within neighborhood
}
\description{
Computes correlation between phenotype and PGS for given neighborhood
}
\details{
This function takes a dataset containing columns for a phenotype and
a PGS for the phenotype as well as a list of indices in the dataset (composing
the "neighborhood") for which to compute the correlation between phenotype
and PGS for.
}
\author{
Nuno R. G. Carvalho: \email{nunocarvalho@gatech.edu}
}
