% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bandPGS_decay.R
\name{get_band_data}
\alias{get_band_data}
\title{get_band_data}
\usage{
get_band_data(data)
}
\arguments{
\item{data}{data table containing all necessary columns and rows}
}
\value{
Returns a tibble with an observation for each bin
}
\description{
makes the band_data tibble used in other functions
}
\details{
Takes a dataset with complete cases and with column names standardized
('dim','pheno','PGS', 'dim_group') and computes per-bin statistics
}
\author{
Nuno R. G. Carvalho: \email{nunocarvalho@gatech.edu}
}
