% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGSdecay.R
\name{get_middle_range}
\alias{get_middle_range}
\title{get_middle_range}
\usage{
get_middle_range(data, ref_window = 0.95)
}
\arguments{
\item{ref_window}{(optional) numeric: proportion of the distribution, centered at the median, to include consider when making the non-leftmost and -rightmost bins. Also affects standardization of portability slope}
}
\value{
Returns a range (length-2 numerical vector)
}
\description{
gets the range of the middle percentile specified
}
\details{
Takes a dataset with complete cases and with column names standardized
('dim','pheno','PGS') and gets the middle range of the dataset
}
\author{
Nuno R. G. Carvalho: \email{nunocarvalho@gatech.edu}
}
