% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NNs.R
\name{get_NNs.FAST}
\alias{get_NNs.FAST}
\title{get_NNs.FAST}
\usage{
get_NNs.FAST(
  data_dims,
  R = NA,
  k = NA,
  mode = "hybrid",
  multiplier = 1,
  coverage = 1,
  seed = NA,
  verbose = FALSE
)
}
\arguments{
\item{data_dims}{data table containing all necessary columns and rows. MUST HAVE NO NA VALUES}

\item{R}{(optional for mode='k') numeric: radius from each point to build a neighborhood from. Only needed if using 'fr' or 'hybrid' mode}

\item{k}{(optional for mode='fr') integer: number of closest neighbors (including self) to build neighborhood from. Only needed if using 'k' or 'hybrid' mode}

\item{mode}{(optional) character: mode of building neighborhoods. One of:
\itemize{
\item 'fr' = fixed-radius mode. Each anchor's neighborhood is composed of every point within a radius R of the anchor point.
\item 'k' = k-nearest mode. Each anchor's neighborhood is composed of the closest k points to the anchor point.
\item 'hybrid' = fr+k mode. Runs fixed-radius mode and then runs k-nearest mode on neighborhoods with sizes less than k. Recommended for reducing noise.
}}

\item{multiplier}{(optional, method='FAST') integer: number of times each point needs to be assigned to a neighborhood before algorithm ends. Bigger = slower = more anchors.}

\item{coverage}{(optional, method='FAST') numeric: proportion of points that need to meet the multiplier parameter before algorithm ends. Bigger = slower = more anchors}

\item{seed}{(optional, method='FAST') numeric: seed used for algorithm}

\item{verbose}{(optional, method='FAST') logical: whether to print out more detailed algorithm-related information}
}
\value{
Returns a nested list, where the i'th element contains a numeric vector of
the i'th point's neighbors (including element i itself) if computed
}
\description{
NN algorithm that builds all possible neighborhoods
}
\details{
This NN algorithm builds a neighborhood for a sample of points
representative of the dimensional space. The algorithm performs the following
loop, starting with zero neighborhoods built:
\enumerate{
\item Randomly sample a point (called the anchor point) in the dataset that is currently part of less than 'multiplier' (default=1) number of neighborhoods
\item Build neighborhood for the anchor point according to 'mode' parameter
\item Update the running list of how many neighborhood each point is part of. Points in the neighborhood built in this iteration of their running count increased by 1.
}
The loop ends once the proportion of points in the dataset that are part of at
least 'multiplier' number of neighborhoods is greater than or equal to
'coverage' (default=1.00)
}
\author{
Nuno R. G. Carvalho: \email{nunocarvalho@gatech.edu}
}
