% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NNs.R
\name{get_NNs.FULL}
\alias{get_NNs.FULL}
\title{get_NNs.FULL}
\usage{
get_NNs.FULL(data_dims, R = NA, k = NA, mode = "hybrid")
}
\arguments{
\item{data_dims}{data table containing all necessary columns and rows. MUST HAVE NO NA VALUES}

\item{R}{(optional for mode='k') numeric: radius from each point to build a neighborhood from. Only needed if using 'fr' or 'hybrid' mode}

\item{k}{(optional for mode='fr') integer: number of closest neighbors (including self) to build neighborhood from. Only needed if using 'k' or 'hybrid' mode}

\item{mode}{(optional) character: mode of building neighborhoods. One of:
\itemize{
\item 'fr' = fixed-radius mode. Each anchor's neighborhood is composed of every point within a radius R of the anchor point.
\item 'k' = k-nearest mode. Each anchor's neighborhood is composed of the closest k points to the anchor point.
\item 'hybrid' = fr+k mode. Runs fixed-radius mode and then runs k-nearest mode on neighborhoods with sizes less than k. Recommended for reducing noise.
}}
}
\value{
Returns a nested list, where the i'th element contains a numeric vector of
the i'th point's neighbors (including element i itself) if computed
}
\description{
NN algorithm that builds all possible neighborhoods
}
\details{
This NN algorithm builds a neighborhood for every point in the data
set (with non-NA data)
}
\author{
Nuno R. G. Carvalho: \email{nunocarvalho@gatech.edu}
}
