% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locPGSacc.FAST.R
\name{locPGSacc.FAST}
\alias{locPGSacc.FAST}
\title{locPGSacc.FAST}
\usage{
locPGSacc.FAST(
  data,
  col_dims,
  col_pheno,
  col_PGS,
  col_PGSacc = "locPGSacc",
  R = -1,
  k = -1,
  mode = "hybrid",
  multiplier = 1,
  coverage = 1,
  seed = NA,
  verbose = FALSE
)
}
\arguments{
\item{data}{data table containing all necessary columns and rows}

\item{col_dims}{character vector: name of the dimension column(s)}

\item{col_pheno}{character: column name of the phenotype of interest}

\item{col_PGS}{character: column name of the polygenic scores for the phenotype of interest}

\item{col_PGSacc}{character: column name of the outputted local PGS accuracy}

\item{R}{numeric: radius from each point to build a neighborhood from. Only needed if using 'fr' or 'hybrid' mode}

\item{k}{integer: number of closest neighbors (including self) to build neighborhood from. Only needed if using 'k' or 'hybrid' mode}

\item{mode}{character: mode of building neighborhoods. One of:
\itemize{
\item 'fr' = fixed-radius mode. Each anchor's neighborhood is composed of every point within a radius R of the anchor point.
\item 'k' = k-nearest mode. Each anchor's neighborhood is composed of the closest k points to the anchor point.
\item 'hybrid' = fr+k mode. Runs fixed-radius mode and then runs k-nearest mode on neighborhoods with sizes less than k. Recommended for reducing noise.
}}

\item{multiplier}{integer: number of times each point needs to be assigned to a neighborhood before algorithm ends. Bigger = slower = more anchors.}

\item{coverage}{numeric: proportion of points that need to meet the multiplier parameter before algorithm ends. Bigger = slower = more anchors}

\item{seed}{numeric: seed used for algorithm}

\item{verbose}{logical: whether to print out more detailed algorithm-related information}
}
\value{
Returns inputted 'data' table but with the following columns appended:
\itemize{
\item 'n_neighbors' = number of neighbors (including self) for that point's neighborhood
\item 'locPGSacc' = correlation between 'col_pheno' and 'col_PGS' columns within that point's neighborhood
}
}
\description{
Fast version of \code{\link[=locPGSacc]{locPGSacc()}} function for large N datasets
}
\details{
For large sample sizes (N>20,000), the full \code{\link[=locPGSacc]{locPGSacc()}} function
may take too long to run and require too much memory. This function solves this
issue by only building neighborhoods for much smaller amount of points called
anchors. This function runs an algorithm to select the optimal set of anchors
to build neighborhoods for such that it is representative of the entire
dimensional space of the data set. Right now only works with quantitative
traits.
}
\author{
Nuno R. G. Carvalho: \email{nunocarvalho@gatech.edu}
}
