% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGSdecay.bs.R
\name{PGSdecay.bs}
\alias{PGSdecay.bs}
\title{PGSdecay.bs}
\usage{
PGSdecay.bs(
  data,
  col_dim,
  col_pheno,
  col_PGS,
  B = 1000,
  i_omit = c(),
  ref_window = 0.95,
  bands = 15,
  return_objects = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{data}{data table containing all necessary columns and rows}

\item{col_dim}{character: column name of the dimensional variable for which to compute bands from}

\item{col_pheno}{character: column name of the phenotype of interest}

\item{col_PGS}{character: column name of the polygenic scores for the phenotype of interest}

\item{B}{(optional) integer: number of bootstrap iterations to perform}

\item{i_omit}{(optional) numeric vector: indices of data rows for which to not consider at all in nearest neighbor algorithm}

\item{ref_window}{(optional) numeric: proportion of the distribution, centered at the median, to include when making bands}

\item{bands}{(optional) integer: number of bands (i.e. sections) to split up the middle of the sample by according to the dimensional variable. An extra band is made on each end to include dispersed samples.}

\item{return_objects}{(optional) logical: whether the function should return the outputs of cor.test() and lm() directly, rather than extracting the most important metrics}

\item{verbose}{(optional) logical: whether bootstrapping iteration status should be reported to the console}
}
\value{
returns a list containing parameters estimated from different bootstrap iterations
\itemize{
\item m: unstandardized slopes
\item m_se: unstandardized slopes' standard errors
\item m_hat: standardized slopes
\item m_hat_se: standardized slopes' standard errors
}
}
\description{
Computes band-based PGS decay through bootstrapping
}
\details{
Works the same as get_bandPGSdecay but employs stratified bootstrapping
to generate multiple estimated regression slopes. Instead of resampling over the
entire sample, resampling is done within each bin to ensure equal representation
across the distance variable of interest. Best used for estimating the
significance of PGS decay along some variable
}
\author{
Nuno R. G. Carvalho: \email{nunocarvalho@gatech.edu}
}
