% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGSdecay.R
\name{PGSdecay}
\alias{PGSdecay}
\title{PGSdecay}
\usage{
PGSdecay(
  data,
  col_dim,
  col_pheno,
  col_PGS,
  i_omit = c(),
  bins = 15,
  return_objects = FALSE
)
}
\arguments{
\item{data}{data table containing all necessary columns and rows}

\item{col_dim}{character: column name of the dimensional variable for which to compute bins from}

\item{col_pheno}{character: column name of the phenotype of interest}

\item{col_PGS}{character: column name of the polygenic scores for the phenotype of interest}

\item{i_omit}{(optional) numeric vector: indices of data rows for which to not consider at all in nearest neighbor algorithm}

\item{bins}{(optional) integer: number of bins to split up the middle of the sample by according to the dimensional variable. The leftmost and rightmost bins will each contain N/bins individuals. The remaining bins will be of varying sample size but equal range}

\item{return_objects}{(optional) logical: whether the function should return the outputs of \code{\link[=cor.test]{cor.test()}} and \code{\link[=lm]{lm()}} directly, rather than extracting the most important metrics}
}
\value{
Returns a nested list with statistics related to PGS decay across bins:
\itemize{
\item bin_data: tibble containing data for each distance bin:
\itemize{
\item bin = range of distance bin
\item min = minimum distance within bin
\item max = max distance within bin
\item N = number of samples in bin
\item r = correlation between phenotype and PGS within bin
\item r_lower = lower r estimate from 95\% confidence interval
\item r_upper = upper r estimate from 95\% confidence interval
\item median = median value of dimensional variable of bin's samples
}
\item lm: computes a linear regression for bin_R2 ~ dim_dist across all bins
\itemize{
\item intercept = intercept of line of best fit
\item m = slope of line of best fit
\item m_se = standard error of the slope of line of best fit
\item p = p-value for slope
\item m_hat = standardized slope; m multiplied by the windowed range of
the dimensional variable and divided by the cor(pheno, PGS) among
reference population (bin with highest r_lower)
\item m_hat_se = standard error of the standardized slope m_hat
}
\item global: computes cor(pheno, PGS) on all samples
\itemize{
\item r = Pearson's correlation coefficient
\item p = p-value
}

}
}
\description{
Returns statistics related to binned PGS accuracy as a function of some continuous/dimensional variable.
}
\details{
Takes a dataset with columns for a phenotype and its PGS as well as a
column for a dimensional variable of interest (e.g. genetic distance) and splits the
sample into equal-length bins. PGS accuracy is calculated within each bin
and the rate of change of accuracy across the dimensional variable is computed.
}
\author{
Nuno R. G. Carvalho: \email{nunocarvalho@gatech.edu}
}
