% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bandPGS_decay.R
\name{plot_bandPGS_decay}
\alias{plot_bandPGS_decay}
\title{plot_bandPGS_decay}
\usage{
plot_bandPGS_decay(
  data,
  col_dim,
  col_pheno,
  col_PGS,
  i_omit = c(),
  window = 0.95,
  bands = 10
)
}
\arguments{
\item{data}{data table containing all necessary columns and rows}

\item{col_dim}{character: column name of the dimension variable for which to compute bands from}

\item{col_pheno}{character: column name of the phenotype of interest}

\item{col_PGS}{character: column name of the polygenic scores for the phenotype of interest}

\item{i_omit}{(optional) numeric vector: indices of data rows for which to not consider at all in nearest neighbor algorithm}

\item{window}{(optional) numeric: proportion of the distribution, centered at the median, to include when making bands}

\item{bands}{(optional) integer: number of bands (i.e. sections) to split up the sample by according to the dimensional variable}
}
\value{
WIP
}
\description{
Computes PGS accuracy along bands of another variable
}
\details{
WIP
}
\author{
Nuno R. G. Carvalho: \email{nunocarvalho@gatech.edu}
}
