% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PGS_decay.R
\name{plot_PGS_decay}
\alias{plot_PGS_decay}
\title{plot_PGS_decay}
\usage{
plot_PGS_decay(
  data,
  col_PGSacc = "locPGSacc",
  col_dist = "dim_dist",
  col_group = NA,
  dist_limits = NA,
  col_pheno = NA,
  col_PGS = NA
)
}
\arguments{
\item{data}{data table containing all necessary columns and rows}

\item{col_PGSacc}{character: column name of the local PGS accuracy}

\item{col_dist}{character: column name of sample's distance in space}

\item{col_group}{(optional) character: column name of sample group assignments}

\item{dist_limits}{(optional) numeric vector: minimum and maximum x-axis limits for distance}

\item{col_pheno}{(optional) character: column name of the phenotype of interest}

\item{col_PGS}{(optional) character: column name of the polygenic scores for the phenotype of interest}
}
\description{
Returns plot of local PGS accuracy as a function of sample distance
}
\details{
Takes a dataset with columns for local PGS accuracy (or any other
desired metric) and sample distance (e.g. genetic PC distance) and plots the
decay of PGS accuracy as a function of distance. Other optional arguments
allow for plotting of PGS accuracy across entire sample and within groups
(e.g. ancestries)
}
\author{
Nuno R. G. Carvalho: \email{nunocarvalho@gatech.edu}
}
