% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalue2text.R
\name{pvalue2text}
\alias{pvalue2text}
\title{pvalue2text}
\usage{
pvalue2text(p_value, round_digits = 3, alpha = 0.05)
}
\arguments{
\item{round_digits}{Integer of total digits to round p-value to}

\item{alpha}{level of significance used so that close values do not get mistaken for being on the border. E.g. p=0.0499999 will never be displayed as p=0.05}

\item{p-value}{numerical p-value to be converted}
}
\value{
Returns a string to be parsed through \code{\link[=plotmath]{plotmath()}} to show up as base10 scientific notation (e.g. 1.23 x 10^-4).
}
\description{
Convert an inputted p-value into a string to be formatted by \code{\link[=plotmath]{plotmath()}}. Uses base10 notation
}
\author{
Nuno R. G. Carvalho: \email{nunocarvalho@gatech.edu}
}
