% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGSdecay.R
\name{bin_dim}
\alias{bin_dim}
\title{bin_dim}
\usage{
bin_dim(data, bins = 15)
}
\arguments{
\item{bins}{(optional) integer: number of bins to split up the middle of the sample by according to the dimensional variable. The leftmost and rightmost bins will each contain N/bins individuals. The remaining bins will be of varying sample size but equal range}
}
\value{
Returns a vector of dim groups to apply directly to the dataset:
}
\description{
splits a complete, renamed dataset by bins along some variable
}
\details{
Takes a dataset with complete cases and with column names standardized
('dim','pheno','PGS') and splits individuals into bins depending on their 'dim'
variable. Internal function.
}
\author{
Nuno R. G. Carvalho: \email{nunocarvalho@gatech.edu}
}
