% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bandPGS_decay.R
\name{get_bandPGS_decay}
\alias{get_bandPGS_decay}
\title{get_bandPGS_decay}
\usage{
get_bandPGS_decay(
  data,
  col_dim,
  col_pheno,
  col_PGS,
  i_omit = c(),
  ref_window = 0.95,
  bands = 15,
  return_objects = FALSE
)
}
\arguments{
\item{data}{data table containing all necessary columns and rows}

\item{col_dim}{character: column name of the dimensional variable for which to compute bands from}

\item{col_pheno}{character: column name of the phenotype of interest}

\item{col_PGS}{character: column name of the polygenic scores for the phenotype of interest}

\item{i_omit}{(optional) numeric vector: indices of data rows for which to not consider at all in nearest neighbor algorithm}

\item{ref_window}{(optional) numeric: proportion of the distribution, centered at the median, to include when making bands}

\item{bands}{(optional) integer: number of bands (i.e. sections) to split up the middle of the sample by according to the dimensional variable. An extra band is made on each end to include dispersed samples.}

\item{return_objects}{(optional) logical: whether the function should return the outputs of cor.test() and lm() directly, rather than extracting the most important metrics}
}
\value{
Returns a nested list with statistics related to banded PGS decay:
\itemize{
\item band_data: tibble containing data for each distance band:
\itemize{
\item band = range of distance band
\item min = minimum distance within band
\item max = max distance within band
\item N = number of samples in band
\item r = correlation between phenotype and PGS within band
\item r_lower = lower r estimate from 95\% confidence interval
\item r_upper = upper r estimate from 95\% confidence interval
\item median = median value of dimensional variable of band's samples
}
\item lm: computes a linear regression for bandPGSacc ~ dim_dist across the bands
\itemize{
\item intercept = intercept of line of best fit
\item m = slope of line of best fit
\item m_se = standard error of the slope of line of best fit
\item p = p-value for slope
\item m_hat = standardized slope; m multiplied by the windowed range of
the dimensional variable and divided by the cor(pheno, PGS) among
reference population (band closest to distance 0)
\item m_hat_se = standard error of the standardized slope m_hat
}
\item global: computes cor(pheno, PGS) on all samples
\itemize{
\item r = Pearson's correlation coefficient
\item p = p-value
}

}
}
\description{
Returns statistics related to banded PGS accuracy as a function of some dimensional variable.
}
\details{
Takes a dataset with columns for a phenotype and its PGS as well as a
column for a dimensional variable of interest (e.g. genetic distance) and splits the
sample into equal-length bands/sections. PGS accuracy is calculated within each
band and the rate of change of accuracy across the dimensional variable is computed.
}
\author{
Nuno R. G. Carvalho: \email{nunocarvalho@gatech.edu}
}
