% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bandPGS_decay.R
\name{plot_bandPGS_decay}
\alias{plot_bandPGS_decay}
\title{plot_bandPGS_decay}
\usage{
plot_bandPGS_decay(
  data,
  col_dim,
  col_pheno,
  col_PGS,
  i_omit = c(),
  ref_window = 0.95,
  bands = 15,
  min_samples = 30,
  fixed_ymin = TRUE
)
}
\arguments{
\item{data}{data table containing all necessary columns and rows}

\item{col_dim}{character: column name of the dimensional variable for which to compute bands from}

\item{col_pheno}{character: column name of the phenotype of interest}

\item{col_PGS}{character: column name of the polygenic scores for the phenotype of interest}

\item{i_omit}{(optional) numeric vector: indices of data rows for which to not consider at all in nearest neighbor algorithm}

\item{ref_window}{(optional) numeric: proportion of the distribution, centered at the median, to include when making bands}

\item{bands}{(optional) integer: number of bands (i.e. sections) to split up the middle of the sample by according to the dimensional variable. An extra band is made on each end to include dispersed samples.}

\item{min_samples}{(optional) integer: minimum number of samples needed in a band to compute PGS accuracy. Must be > 2}

\item{fixed_ymin}{(optional) logical: whether the y-axis should start at 0 (default) or just be the minimum accuracy}
}
\value{
Plot of the results of \code{\link[=get_bandPGS_decay]{get_bandPGS_decay()}}.
}
\description{
Computes PGS accuracy along bands of another variable
}
\details{
Plots the decay of PGS accuracy along some dimensional variables by
cutting up the dimensional variable into bands/sections. More details in
\code{\link[=get_bandPGS_decay]{get_bandPGS_decay()}}.
}
\author{
Nuno R. G. Carvalho: \email{nunocarvalho@gatech.edu}
}
