% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_locPGS_decay.R
\name{plot_locPGS_decay}
\alias{plot_locPGS_decay}
\title{plot_locPGS_decay}
\usage{
plot_locPGS_decay(
  data,
  col_PGSacc = "locPGSacc",
  col_dist = "dim_dist",
  col_group = NA,
  dist_limits = NA,
  col_pheno = NA,
  col_PGS = NA
)
}
\description{
Plots local PGS accuracy as a function of sample distance
}
\details{
Takes a dataset with columns for local PGS accuracy (or any other
desired metric) and sample distance (e.g. genetic PC distance) and plots the
decay of PGS accuracy as a function of distance. Other optional arguments
allow for plotting of PGS accuracy across entire sample and within groups
(e.g. ancestries). Annotated statistics are computed from \code{\link[=get_locPGS_decay]{get_locPGS_decay()}}.
}
\author{
Nuno R. G. Carvalho: \email{nunocarvalho@gatech.edu}
}
